// src/types/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      product_variants: {
        Row: {
          color_id: string | null
          id: string
          is_deleted: boolean
          product_id: string
          retail_price: number
          size_id: string | null
          sku: string | null
          stock_quantity: number
          low_stock_threshold: number
          wholesale_price: number
        }
        Insert: {
          color_id?: string | null
          id?: string
          is_deleted?: boolean
          product_id: string
          retail_price?: number
          size_id?: string | null
          sku?: string | null
          stock_quantity?: number
          low_stock_threshold?: number
          wholesale_price?: number
        }
        Update: {
          color_id?: string | null
          id?: string
          is_deleted?: boolean
          product_id?: string
          retail_price?: number
          size_id?: string | null
          sku?: string | null
          stock_quantity?: number
          low_stock_threshold?: number
          wholesale_price?: number
        }
        Relationships: [
          {
            foreignKeyName: "product_variants_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          }
        ]
      }
      products: {
        Row: {
          category_id: string | null
          created_at: string
          description: string | null
          id: string
          is_deleted: boolean
          name: string
          shop_id: string
        }
        Insert: {
          category_id?: string | null
          created_at?: string
          description?: string | null
          id?: string
          is_deleted?: boolean
          name: string
          shop_id: string
        }
        Update: {
          category_id?: string | null
          created_at?: string
          description?: string | null
          id?: string
          is_deleted?: boolean
          name?: string
          shop_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "products_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "products_shop_id_fkey"
            columns: ["shop_id"]
            isOneToOne: false
            referencedRelation: "shops"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
    }
// ...
  }
}